fwupdate: Add check to skip removing pldm directory in /tmp/images when pldmd restarts.
For non-pldm devices images will be copied to /tmp/images/pldm
directory and it's created by item updater. When pldmd process
restarts it removes the files and folder in /tmp/images. Since
these directories are no longer valid image copy fails, which
results in update failure.

This patch adds check to skip removing pldm directory in
/tmp/images when pldmd restarts.

Unit Tests:
  1. Verify pldm directory is not removed when pldm restarts
  2. Verify other directories are deleted except pldm
  3. Firmware update sanity with and without pldm restart

Fixes nvbug https://nvbugs/3912340

+++ fw-update/update_manager.cpp
@@ -38,7 +38,17 @@ UpdateManager::UpdateManager(
     {
         for (const auto& entry :
              std::filesystem::directory_iterator("/tmp/images"))
-            std::filesystem::remove_all(entry.path());
+        {
+            if (entry.is_directory() && entry.path() == "/tmp/images/pldm")
+            {
+                // skip removing pldm directory which contains non-pldm image
+                // directories.
+            }
+            else
+            {
+                std::filesystem::remove_all(entry.path());
+            }
+        }
     }
     progressTimer = nullptr;
 }
